Fixed libxslt to check if python is actually available
Note:

It seems there are issues connected to the PYTHON environment variable,
leaking into the sandbox if python is installed on the system the target
is build on. Nevertheless it seems to be not a big problem it might be
safer to add this check!
The same problem also appears while building the libxml package!