coreutils: fixed duplication of binaries by using install instead of ./ginstall for installing at native build stages
Note:

When not cross-compiling the just-built ./ginstall is called for installing
the binaries. Because this circumvents the OpenSDE install wrapper, defined
install wrapper filters will not have any effect in "native" build stages.
The coreutils package is heavily relying on install wrapper filters for
changing the install location of certain binaries from '/usr/bin' to '/bin'
to be FHS compliant.
Without this fix those binaries will get installed at both locations,
because on cross-stage the filter is working properly.