cups: fixed build by setting NO-LIBTOOL-FIX flag in the .desc file (see note)
Note:
If the NO-LIBTOOL-FIX flag is not set we export the LIBTOOL environment
variable, especially for cross-build helping to pickup the right libtool.
The Cups build environment doesn't officially support libtool based builds
and they clearly warn to use it to build cups, but their Makefiles are not
properly prepared for the case the LIBTOOL env var is exported.