firefox: fixed to not use system's libpng (see note)
Note:

Firefox needs a PNG library that comes with APNG (Animated Portable Network
Graphics) support, which is not included in our standard libpng.

The internal version Mozilla is using is a patched libpng which has APNG
support added, and the authors also provide patches for stock libpng, but
it seems that libpng's authors have no real interest in integrating it.

Patches and further info: http://littlesvr.ca/apng/