crypto++: Added crypto++ 5.6.1 - A free C++ class library of cryptographic schemes (see note)
Note:

The package includes a patch which changes the makefile:

This patch is changing the crypto++ makefile to build both, a static
and a shared version of the library, but not the supplied test
application.

TODO:

There is quite some room for improvements inside of the makefile, so
it would be good to have an more "inteligent" way to select what kind
of library should be build (i.e. static and shared, static-only,
shared-only).
Besides that the 'install' target is very poorly implemented and should
be improved to consider the variants mentioned above!