runit/havaged: changed run script template to add '-w 3072' (see note)
Note:

The default size of the linux kernel entropy pool is 4096 bits, and the default
'write_wakup_threshold' - defining the number of bits of entropy below which we
wake up processes that do a select() or poll() for write access to /dev/random
- is set to 128, which is not sufficient for machines that consume a lots of
entropy.
The 'write_wakup_threshold' can be set manually via sysfs in
'/proc/sys/kernel/random/write_wakeup_threshold' but haveged is able to set it
by itself through the '-w' option.