openvpn: improved to build plugins (see note)
Note:

This commit is improving the openvpn package to build the included
plugins if possible (currently only the pam authentication plugin
if pam is installed), but also introducing the 'plugins_enabled'
variable within the openvpn.conf file which can be used to define
a space seperated list of plugins which should be build additionally.

This will offer a convinient way for targets which patch in custom
plugins to enable them within the build process of the package.
The only thing to do is to append the desired plugin to then list
stored within the $plugins_enabled variable.

Example: Adding 'myplugin' within a target specific 'openvpn.conf'
--------------------------------------------------------------------
var_append plugins_enabled ' ' "myplugin"
--------------------------------------------------------------------