lprof: moved into attic because it is qt3 only (see note!)
Note:

The last official release of lprof is qt3 only, and development seems to be
stalled. The nearly 3 years old cvs version is using qt4 but is relying on
qt-assistant which got deprecated with qt 4.7.
Thus the only way to re-activate this package is to package the qt-assistant
compatibility package and to use the CVS version of lprof!