libchipcard3: renamed to libchipcard (see note!)
Note:

Historicaly we have 3 different libchipcard packages which is not really
needed anymore, so I renamed _our_ latest package (libchipcard3) to
libchipcard, completely replacing the very ancient version we had inside
of that package.
libchipcard2 will be dropped as well but not within this commit, and the
the remaining libchipcard package updated to a recent upstream version.