metamail: Updated (2.7 -> 2.7-54) switching to debian's metamail version
Note:

metamail is a kind of dinosaur of a package, and because there seems to be
no official development going on upstream, and the formerly download
location seems to be pretty dead, I switched to the source tarball provided
by debian which itself includes more tools. Besides that I removed all
patches we had locally and incorporated the official debian patchset.