kbarcode: moved into attic (see note!)
Note:

kbarcode is not fully ported to KDE4 yet, but the author is providing a less
powerful kde4 version called kbarcode4-light, which is similar to KBarcode for
KDE3, but less powerful. KBarcode4-light allows only to create single barcodes.
As the port for KBarcode is not yet ready and
no one knows when it will be ready, KBarcode4-light provides basic barcode
generation for KDE. An additional benefit is that it provides an easy to use
interface for Barcode Writer in Pure Postscript.