xaralx: moved into attic (see note!)
Seems that since MAGIX AG bought Xara Group Ltd. there is no budget for
developing the OSS vararian xaralx anymore, and according to Wikipedia:

http://en.wikipedia.org/wiki/Xara_Xtreme#Versions_for_Linux

"The progress of the Linux version has stalled and the program's website has
not been updated since August 2006. No commits to SVN source code repository
has been happening for same period of time.
A new project, xarino, was created in December 2008[5], which provides access
to the project's source code. A new mailing list was established in May 2009"

http://code.google.com/p/xarino/

But xarino seems to be stalled as well (last commit to svn 2009/01/25) so it
looks xaralx/xarino will rest in peace.