gstreamer: changed build priority (155.400 -> 127.000) and to source glib-conf.in instead of gtk+-conf.in (see note!)
Note:

Besides the usual toolchain components we only need glib and libxml2 to
build the gstreamer core, thus it doesn't have to be build after gtk+
and there is no need to source the gtk+-conf.in.
To make gstreamer available to more packages the build priority has to
be changed to something right after glib and libxml2.