webkit: renamed webkitgtk (see note!)
Note:

Because our webkit package is build the gtk port by default we agreed in
renaming the package to webkitgtk and use the offical stable releases from
the WebKitGTK+ project (http://webkitgtk.org).

The Qt port of webkit is included within qt >= 4.7 so there is no real
need to build that port from the nightly snapshots we are currently unsing!

For other ports like the EFL port we have to decide later how we are going
to handle those cases!