sqlite: Updated (3.7.3 -> 3.7.6.2) (see note!)
Note:

Beginning with SQLite version 3.7.4 the naming and versioning scheme changed
again!

This update also replaces the "pure" amalgamation source tarball in favour of
the "autoconf" version, which is a tarball containing the amalgamation for
SQLite 3.7.6.2 together with an configure script and makefile for building it.
This tarball also contains in the "tea" subdirectory a separate configure
script and makefile compatible with the Tcl Extension Architecture (TEA) for
building the TCL bindings to SQLite.