libvserver: moved into attic
Note:

There seems to be no active development of libvserver anymore, but to be
sure we should ask Benedikt (hollow) if he will continue to work on it
some day. Actually th libvserver package should build fine but it conflicts
with the util-vserver package, because it is using the same naming scheme
for the library and development headers.