glibc: reintroduced handling of internal libidn support (see note!)
Note:

This partially reverts commit cde2856ca0a718e437124ee17e545521104d418e
because I wrongly assumed that the IDN support will be enabled by default,
which is not the case! As of glibc 2.11 only the sources of the glibc
libidn addon are included, but the addon itself is *NOT* enabled by
default. Thus I restored the old behavior besides the part for unpacking
the libidn addon sources.