glibc: added patch to reexport RPC interface
Note:

Thanks to Mr. Drepper the sunrpc interface was removed but third-party
application are still depending on it, thus we are reexporting the RPC
interface.
This patch was made by Andreas Schwab and was taken from fedora/master
branch.
For me it is a mysteria why one person from Redhat is first removing
code from the official glibc release and another one readding parts of
it to a Redhat maintained branch! Thank you Redhat!