ncurses: Updated (5.9-20110423 -> 5.9-20120107)
Important

the 'pkg_config_sysroot.patch' was disabled because it didn't applied
correctly but we need to fix that issue again.
Besides that we have to fix some other pkgconfig related parts within
configure to honor $PKG_CONFIG_SYSROOT if it is set.
One thing I have observed is that it currently checks the pkgconfig
directories of the host for deciding to install .pc files when doing
a cross build.