embryo: fixed license information (GPL -> BSD)
Note:

Actually it is a BSD 2-clause license, but as stated in the included COPYRIGHT
file large parts are licensed under a different license.

-----COPYRIGHT file snip -------------------------------------------------

Large parts of this source come under the following license from the
original Small (renamed to Pawn afterwards, but was named Small when
Embryo split off). See the source files that are clearly marked as below:

Copyright (c) ITB CompuPhase, 1997-2003

This software is provided "as-is", without any express or implied warranty.
In no event will the authors be held liable for any damages arising from
the use of this software.

Permission is granted to anyone to use this software for any purpose,
including commercial applications, and to alter it and redistribute it
freely, subject to the following restrictions:

 1.  The origin of this software must not be misrepresented; you must not
     claim that you wrote the original software. If you use this software in
     a product, an acknowledgment in the product documentation would be
     appreciated but is not required.
 2.  Altered source versions must be plainly marked as such, and must not be
     misrepresented as being the original software.
 3.  This notice may not be removed or altered from any source distribution.

-----COPYRIGHT file snap -------------------------------------------------