libpng: changed to build at toolchain stage if requested
Note:

This is needed as we need a fairly complete EFL stack to build edje. We need
edje_cc capable to run on the build host to "compile" default edje theme files
which are shipped with most EFL based applications.
libpng itself is needed to get png loaders build within the EFL stack, which
is needed to handle png graphics when "compiling" the default theme using
edje_cc.