linux-header: improved to remove unneeded hidden files
Note:

By default the 'headers_install' make target installs hidden files within
the destination directories called '..install.cmd' and '.install' and the
'headers_check' make target is doing it similar, by installing files like
'..check.cmd' and '.check'.
Those files seem to be only interesting for kernel developers and are not
needed within the package, thus we are going to remove the files.