mdadm: fixed cross-compiling when toolchain is only build at stage 0
Cross-compiling mdadm fails in the case binutils was only build in stage 0.
The ansidecl.h is is only installed if the --enable-shared configure option
is given when building binutils which is not the case at toolchain stage
(aka stage 0). But even if we provide the --enable-shared configure option
at toolchain stage, the headers are installed at place which is not known
when building packages at stage 1.