radiusclient-ng: renamed/updated to freeradius-client (0.5.6 -> 1.1.6)
From http://freeradius.org/freeradius-client/

History

Radiusclient used to live at http://www.cityline.net/~lf/radius/ but seems to
be no longer maintained (Tarballs are still available from
ftp://ftp.cityline.net/pub/radiusclient/). It does however still ship with some
Linux distributions including SUSE Linux.

A fork of the original radiusclient called radiusclient-ng is used as the basis
of the RADIUS support in SER and OpenSER.

In late 2006 it was decided that the FreeRADIUS Project should adopt the latest
code from radiusclient-ng cvs as the basis of a new FreeRADIUS client package.
The intention is to roll some of the client functionality of the FreeRADIUS
server into this separate client package and to make the FreeRADIUS suite a one
stop shop for all people wishing to use and develop RADIUS related
applications.

The first version released by the FreeRADIUS project is 1.1.5. This version
fixes an MD5 related endian issue and introduces a client configuration
framework to the library which lends itself to allowing embedded radius client
functionality. Development of the codebase is ongoing.