net-tools: Updated (1.60 -> 1.60.20121017-git8217b88b)
Note:

There wasn't a release net-tools for 11 years now, but development it seems
that upstream development is going on at http://net-tools.sourceforge.net,
thus we are going to use a recent git snapshot now, for people who still
like to use ifconfig and the other historical commands included within the
net-tools package!

Please also note that this commit is removing the ether-wake command which
was patched in before by using a patch provided by linuxfromscratch.

http://www.linuxfromscratch.org/patches/downloads/net-tools/net-tools-1.60-ether_wake-1.patch