ulogd2: fixed to not use the FILE struct as marker to check if the user requested to use syslog
Do not use the FILE struct as marker to check if the user requested to use
syslog.

In musl the FILE struct is opaque, thus its storage size is not known to
the application.
This means that when using musl it has to be use as a pointer like it was
intended to be used.