musl: fixed to always install libc headers
This changes the behavior which was to install headers only at stage 0,
which was sufficient to build a proper toolchain and to cross-compile,
but as the headers have not been installed at stage 1 we have not header
files within the flist of the musl package.
Now we are always calling the 'install-headers' target when calling make
install, but we need to 'touch' every header file within the sources to
be able to install the headers even if already existing in the build
sandbox.