gcc: fixed to not inject -rpath when cross-compiling
Resulting binaries like libstdc++.so will end up having an rpath with an
absolute path to the build sandbox set in the elf headers.
This issue was discovered when trying to run a cross-compiled lftp which
was cross-compiled against musl libc and segfaulted when being executed.
Removing the explicit rpath seems to solve the runtime issue because lftp
is not segfaulting anymore and works as expected.